refactor, warnings suppressed, items imgs changed
