+++ src/components/user/UserPage.css
@@ -2,6 +2,7 @@
     height: 100%;
     display: flex;
     padding: 0px 15%;  
+    
 }
 
 .user-page-navigator {
