+++ src/components/user/UserView.js
@@ -1,4 +1,3 @@
-import { Typography } from '@material-ui/core';
 import React from 'react';
 import { connect } from 'react-redux';
 import GuestPanel from './GuestPanel';
@@ -15,7 +14,6 @@ const UserView = props => {
     console.log('user view', props)
     return (
         <section id='user-view' className='user-view'>
-            {/* <Typography>logged in: {props.isLoggedIn ? 'yes' : 'no'}</Typography> */}
             {/* <GuestPanel /> */}
             <UserPage />
         </section>
