+++ src/redux/actions/productActions.js
@@ -35,7 +35,7 @@ export const fetchProducts = url => dispatch => {
 
     setTimeout( () => {
         dispatch(receiveTest());
-    }, 1000);
+    }, 100);
     /*return axios({
         url: url,
         timeout: 10000,
